requirements.txt
# Core ontology profile generation uses only the Python standard library.

# Document extraction
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-docx>=1.1.0
pytesseract>=0.3.10

# Optional OCR preprocessing for future extensions
opencv-python>=4.8.0
numpy>=1.24.0
